move react-switch
