+++ packages/react-components/react-switch/.babelrc.json
