+++ packages/react-components/react-switch/.eslintrc.json
