+++ packages/react-components/react-switch/LICENSE
