+++ packages/react-components/react-switch/MIGRATION.md
