+++ packages/react-components/react-switch/Spec.md
