+++ packages/react-components/react-switch/bundle-size/Switch.fixture.js
