+++ packages/react-components/react-switch/config/api-extractor.json
