+++ packages/react-components/react-switch/config/tests.js
