+++ packages/react-components/react-switch/etc/react-switch.api.md
@@ -3,7 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-
 import type { ComponentProps } from '@fluentui/react-utilities';
 import type { ComponentState } from '@fluentui/react-utilities';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
@@ -26,21 +25,22 @@ export const switchClassNames: SlotClassNames<SwitchSlots>;
 
 // @public (undocumented)
 export type SwitchOnChangeData = {
-    checked: boolean;
+  checked: boolean;
 };
 
 // @public
-export type SwitchProps = Omit<ComponentProps<Partial<SwitchSlots>, 'input'>, 'onChange'> & Partial<SwitchCommons> & {
+export type SwitchProps = Omit<ComponentProps<Partial<SwitchSlots>, 'input'>, 'onChange'> &
+  Partial<SwitchCommons> & {
     defaultChecked?: boolean;
     onChange?: (ev: React_2.ChangeEvent<HTMLInputElement>, data: SwitchOnChangeData) => void;
-};
+  };
 
 // @public (undocumented)
 export type SwitchSlots = {
-    root: NonNullable<Slot<'div'>>;
-    indicator: NonNullable<Slot<'div'>>;
-    input: NonNullable<Slot<'input'>>;
-    label?: Slot<typeof Label>;
+  root: NonNullable<Slot<'div'>>;
+  indicator: NonNullable<Slot<'div'>>;
+  input: NonNullable<Slot<'input'>>;
+  label?: Slot<typeof Label>;
 };
 
 // @public
@@ -53,5 +53,4 @@ export const useSwitch_unstable: (props: SwitchProps, ref: React_2.Ref<HTMLInput
 export const useSwitchStyles_unstable: (state: SwitchState) => SwitchState;
 
 // (No @packageDocumentation comment for this package)
-
 ```
