+++ packages/react-components/react-switch/src/common/isConformant.ts
