+++ packages/react-components/react-switch/src/components/Switch/Switch.test.tsx
