+++ packages/react-components/react-switch/src/components/Switch/renderSwitch.tsx
