+++ packages/react-components/react-switch/src/components/Switch/useSwitchStyles.ts
