+++ packages/react-components/react-switch/src/stories/SwitchBestPractices.md
