+++ packages/react-components/react-switch/src/stories/SwitchChecked.stories.tsx
