+++ packages/react-components/react-switch/tsconfig.lib.json
