+++ packages/react-components/react-switch/tsconfig.spec.json
